+++ example/.umirc.ts
@@ -1,8 +1,11 @@
 import { defineConfig } from 'umi';
 
 export default defineConfig({
-  presets: [require.resolve('../packages/preset-react/lib')],
-  // plugins: [require.resolve('../packages/plugin-webpack-5/lib')],
+  presets: [
+    require.resolve('../packages/preset-react/lib'),
+    // require.resolve('../packages/plugin-esbuild'),
+  ],
+  plugins: [require.resolve('../packages/plugin-esbuild/lib')],
   routes: [
     {
       name: 'model 测试',
@@ -26,6 +29,11 @@ export default defineConfig({
       path: '/request',
       component: './request',
       menu: false,
+    },
+    {
+      name: 'locale 测试',
+      path: '/plugin-locale',
+      component: './plugin-locale',
     },
     {
       name: '首页',
