+++ example/pages/plugin-locale.css
